backend/requirements.txt
annotated-types==0.7.0
bidict==0.23.1
blinker==1.9.0
cachecontrol==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
email-validator==2.2.0
firebase-admin==6.7.0
flask==3.1.0
flask-cors==5.0.1
flask-socketio==5.5.1
flask-sqlalchemy==3.1.1
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jsonify==0.5
markupsafe==3.0.2
msgpack==1.1.0
pika==1.3.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pyjwt==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
rsa==4.9
simple-websocket==1.1.0
sqlalchemy==2.0.39
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
werkzeug==3.1.3
wsproto==1.2.0
